Add DSQL migration plan for CASCADE/CHECK constraints (#25)
* Add DSQL migration plan for CASCADE/CHECK constraints

Documents the plan to move all ON DELETE CASCADE and CHECK constraints
from the database layer into application code for AWS Aurora DSQL
compatibility.

Key contents:
- Dependency graph showing deletion order for all tables
- Affected tables and current constraints
- Implementation plan with code examples for all delete functions
- Testing strategy for validation and cascade behavior
- DSQL-specific considerations (row limits, OCC)
- Rollback plan


* Implement application-level CASCADE/CHECK for DSQL compatibility

Move all ON DELETE CASCADE and CHECK constraint logic from the database
layer into application code for AWS Aurora DSQL compatibility.

Updated delete functions with transactional cascades:
- delete_user(): Full cascade across 10+ child tables
- delete_authenticator(): Cascade sessions, clear device_auth refs
- delete_oauth_client(): Cascade secrets and usage events
- delete_scim_token(): SET NULL for audit log references
- delete_scim_group(): Cascade group memberships
- delete_expired_device_auth_requests(): Cascade oidc_states
- delete_organization(): New function with full cascade

Added migration 024_dsql_compatibility.sql that recreates all affected
tables without CASCADE/CHECK constraints. Basic FK constraints are
preserved for data integrity during the SQLite transition period.


---------

+++ crates/vouch-server/src/handlers/browser_login.rs
@@ -421,8 +421,8 @@ pub async fn browser_login_complete(
         verification_result.user_verified
     );
 
-    // Update counter in database
-    let new_counter = i64::from(verification_result.counter);
+    // Update counter in database (WebAuthn counter is u32, stored as i32)
+    let new_counter = verification_result.counter as i32;
     db::update_authenticator_counter(&state.db, &authenticator.id, new_counter)
         .await
         .map_err(|e| {
